kerneldetection/requirements.txt
absl-py==0.11.0
astor==0.8.1
cached-property==1.5.2
decorator==4.4.2
et-xmlfile==1.0.1
gast==0.2.2
google-pasta==0.2.0
grpcio==1.33.2
h5py==3.1.0
importlib-metadata==3.1.0
jdcal==1.4.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
Markdown==3.3.3
networkx==2.5
numpy==1.19.4
openpyxl==3.0.5
opt-einsum==3.3.0
pandas==1.1.4
protobuf==3.14.0
python-dateutil==2.8.1
pytz==2020.4
six==1.15.0
tensorboard==1.15.0
tensorflow==1.15.0
tensorflow-estimator==1.15.1
termcolor==1.1.0
Werkzeug==1.0.1
wrapt==1.12.1
XlsxWriter==1.3.7
zipp==3.4.0
onnx==1.9.0
networkx==2.5.1
